# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProviderStatus {
  configured: boolean;
  provider: string;
  model: string;
}

export interface TrendingVerdict {
  query: string;
  category: string;
}

export interface VerdictInput {
  /**
     * @minLength 2
     * @maxLength 180
     */
  query: string;
}

export type VerdictSourceStatus = typeof VerdictSourceStatus[keyof typeof VerdictSourceStatus];


export const VerdictSourceStatus = {
  available: 'available',
  unavailable: 'unavailable',
} as const;

export interface ReviewSource {
  title: string;
  publisher: string;
  url: string;
  /** @nullable */
  publishedAt: string | null;
}

export interface Verdict {
  query: string;
  title: string;
  summary: string;
  verdict: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  recommendation: string;
  pitfalls: string[];
  evidence: string[];
  popularReviews: string[];
  sources: ReviewSource[];
  sourceStatus: VerdictSourceStatus;
  sourceStatusMessage: string;
  generatedAt: string;
  model: string;
}

export interface ApiError {
  error: string;
  code?: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export type Plan = typeof Plan[keyof typeof Plan];


export const Plan = {
  free: 'free',
  plus: 'plus',
  unlimited: 'unlimited',
} as const;

export interface QuotaStatus {
  plan: Plan;
  /** @minimum 0 */
  used: number;
  /**
     * @minimum 0
     * @nullable
     */
  limit: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  remaining: number | null;
  periodStart: string;
  /** @nullable */
  resetsAt: string | null;
}

export type ActivatePlanRequestPlan = typeof ActivatePlanRequestPlan[keyof typeof ActivatePlanRequestPlan];


export const ActivatePlanRequestPlan = {
  plus: 'plus',
  unlimited: 'unlimited',
} as const;

export interface ActivatePlanRequest {
  plan: ActivatePlanRequestPlan;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};