# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TruthRouter AI API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: verdicts
  - name: provider
  - name: Auth
    description: Browser authentication endpoints.
  - name: account
  - name: admin
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /provider-status:
    get:
      operationId: getProviderStatus
      tags: [provider]
      summary: Return the configured AI provider and model
      responses:
        "200":
          description: Provider status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderStatus"
  /trending-verdicts:
    get:
      operationId: getTrendingVerdicts
      tags: [verdicts]
      summary: Return popular example verdict queries
      responses:
        "200":
          description: Trending verdicts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendingVerdict"
  /verdict:
    post:
      operationId: createVerdict
      tags: [verdicts]
      summary: Generate an honest product verdict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerdictInput"
      responses:
        "200":
          description: Generated verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verdict"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Too many verdict requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the identity provider.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after login.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the browser session
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after logout.
  /account/quota:
    get:
      tags: [account]
      operationId: getAccountQuota
      summary: Get the signed-in account's current review quota
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Current quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaStatus"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /admin/users/{userId}/plan:
    post:
      tags: [admin]
      operationId: activateUserPlan
      summary: Activate a paid plan after subscription confirmation
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivatePlanRequest"
      responses:
        "200":
          description: Plan activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaStatus"
        "400":
          description: Invalid plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Administrator access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ProviderStatus:
      type: object
      required: [configured, provider, model]
      properties:
        configured:
          type: boolean
        provider:
          type: string
        model:
          type: string
    TrendingVerdict:
      type: object
      required: [query, category]
      properties:
        query:
          type: string
        category:
          type: string
    VerdictInput:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 2
          maxLength: 180
    Verdict:
      type: object
      required:
        [query, title, summary, verdict, confidence, recommendation, pitfalls, evidence, popularReviews, sources, sourceStatus, sourceStatusMessage, generatedAt, model]
      properties:
        query:
          type: string
        title:
          type: string
        summary:
          type: string
        verdict:
          type: string
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        recommendation:
          type: string
        pitfalls:
          type: array
          items:
            type: string
        evidence:
          type: array
          items:
            type: string
        popularReviews:
          type: array
          items:
            type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/ReviewSource"
        sourceStatus:
          type: string
          enum: [available, unavailable]
        sourceStatusMessage:
          type: string
        generatedAt:
          type: string
          format: date-time
        model:
          type: string
    ReviewSource:
      type: object
      required: [title, publisher, url, publishedAt]
      properties:
        title:
          type: string
        publisher:
          type: string
        url:
          type: string
          format: uri
        publishedAt:
          type: [string, "null"]
          format: date-time
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    Plan:
      type: string
      enum: [free, plus, unlimited]
    QuotaStatus:
      type: object
      required: [plan, used, limit, remaining, periodStart, resetsAt]
      properties:
        plan:
          $ref: "#/components/schemas/Plan"
        used:
          type: integer
          minimum: 0
        limit:
          type: [integer, "null"]
          minimum: 0
        remaining:
          type: [integer, "null"]
          minimum: 0
        periodStart:
          type: string
          format: date
        resetsAt:
          type: [string, "null"]
          format: date-time
    ActivatePlanRequest:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [plus, unlimited]