# lib/api-zod/src/generated/types/authUserEnvelope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthUserEnvelope {
  user: AuthUser | null;
}