# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */

export type Plan = typeof Plan[keyof typeof Plan];


export const Plan = {
  free: 'free',
  plus: 'plus',
  unlimited: 'unlimited',
} as const;