# lib/api-zod/src/generated/types/providerStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ProviderStatus {
  configured: boolean;
  provider: string;
  model: string;
}