# lib/api-zod/src/generated/types/quotaStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Plan } from './plan';

export interface QuotaStatus {
  plan: Plan;
  /** @minimum 0 */
  used: number;
  /**
     * @minimum 0
     * @nullable
     */
  limit: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  remaining: number | null;
  periodStart: Date;
  /** @nullable */
  resetsAt: Date | null;
}