# lib/api-zod/src/generated/types/verdict.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TruthRouter AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewSource } from './reviewSource';
import type { VerdictSourceStatus } from './verdictSourceStatus';

export interface Verdict {
  query: string;
  title: string;
  summary: string;
  verdict: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  recommendation: string;
  pitfalls: string[];
  evidence: string[];
  popularReviews: string[];
  sources: ReviewSource[];
  sourceStatus: VerdictSourceStatus;
  sourceStatusMessage: string;
  generatedAt: Date;
  model: string;
}